requirements.txt
openai>=1.0,<2.0
python-dotenv>=1.0,<2.0
streamlit>=1.36,<2.0
pytest>=7.0,<8.0
pytest-cov>=4.0,<5.0
pytest-mock>=3.10,<4.0
pre-commit>=3.0,<4.0
